chore(Geometry/Euclidean/Triangle): refactor proof of sum of angles in a triangle (#29299)
This PR refactors the proof that the sum of angles in a triangle is pi. This is done by proving it from the theorem that `angle x y = angle x (x + y) + angle y (x + y)`. We also use the new `field_simp` tactic to shorten the proofs. The new proof is quite a bit shorter.